refactor: 도메인 내부 DTO 증가로 파일 분류 및 폴더 이동
