+++ src/main/java/team03/mopl/domain/watchroom/dto/participant/ParticipantDto.java
@@ -1,4 +1,4 @@
-package team03.mopl.domain.watchroom.dto;
+package team03.mopl.domain.watchroom.dto.participant;
 
 import lombok.Builder;
