+++ src/main/java/team03/mopl/domain/watchroom/dto/participant/ParticipantsInfoDto.java
@@ -1,4 +1,4 @@
-package team03.mopl.domain.watchroom.dto;
+package team03.mopl.domain.watchroom.dto.participant;
 
 import java.util.List;
 import lombok.Builder;
