+++ src/main/java/team03/mopl/domain/watchroom/dto/video/VideoControlRequest.java
@@ -1,4 +1,4 @@
-package team03.mopl.domain.watchroom.dto;
+package team03.mopl.domain.watchroom.dto.video;
 
 import team03.mopl.domain.watchroom.entity.VideoControlAction;
